requirements.txt
apistar==0.7.2
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
dj-rest-auth==2.1.2
Django==3.1.4
django-allauth==0.44.0
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-solo==1.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
importlab==0.5.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
Markdown==3.3.3
MarkupSafe==1.1.1
networkx==2.5
ninja==1.10.0.post2
oauthlib==3.1.0
packaging==20.8
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
six==1.15.0
South==1.0.2
sqlparse==0.4.1
swapper==1.1.2.post1
typed-ast==1.4.1
typesystem==0.2.4
uritemplate==3.0.1
urllib3==1.26.2
